requirements.txt
pip                 23.2.1
asyncio-nats-client 0.11.5
fastapi             0.101.0
uvicorn[standard]   0.23.2